Add developer responsiveness badges (#774)
* Add developer responsiveness badges

* Use rounded response rate for badges

* Move developer response rate into Developer model.

The developer response rate is stored in the `developers` table. This
value will be updated a configurable time after a new conversation is
intitiated by a Business.
This commit includes a migration to add the db column.
This commit includes a rake task (rake
backfills:developer_response_rate) to be run after the migration.

* rename 'time allowed' to 'grace period'

* Consolidate High Response Rate badges into a single badge.

 If a developer has a 90%+ response rate show a badge that says "High Response Rate".

* Move developer response rate updating to Messages::Notifications

* Remove migration check from backfills:developer_response_rate task

Let the job fail on its own when the migration hasn't been run.

* Improve developer response rate testing

Simplify testing now that we are storing the response rate as an actual percentage.
Additional tests to check that the grace period for replying to business contacts is being honored.

* Lint files

* Update Changelog

* Add a note to change persistance method later.

We are using update_column to avoid changing the updated_at
timestamp on the developer record which would trigger the rendering of a
'Recently Active' badge on the developer profile.
This should be changed to update! once we have implemented cacheing of 
the
badges in their own model.

* i18n-tasks normalize

* fix issue with seeds trying to schedule background jobs

We set the queue adapter to :async for UpdateDeveloperResponseRateJob to avoid a NotImplementedError when seeding Conversations and
set the config.developer_response_grace_period to zero so that the jobs are run immediately.

* Run UpdateDeveloperResponseRateJob after seeding Conversations

This ensures that the correct developer badges are seen in a freshly seeded development environment

* Fix developer response rate logic

This fix will ensure any conversations replied to within the grace period are also included
in the response rate calculation.

* Update CHANGELOG.md

Co-authored-by: Joe Masilotti <joe@masilotti.com>

* Remove unecessary calls to UpdateDeveloperResponseRateJob in seeds.

* Remove database delivery for business and developer welcome notifications (#790)

* Disable database delivery for business and developer welcome emails

* Remove @notification from welcome mailer methods

* RailsDevs link to root_url

* Update CHANGELOG.md

* use subject directly from translations

* Point out purpose of PR

* Clean up translations and formatting

* Remove notifications and just send an email

* Enable business welcome emails!

* Match order of BusinessMailer

* No more notification = no more recipient

---------

Co-authored-by: Joe Masilotti <joe@masilotti.com>

* Tidy up tests

* Add tests for Conversation#developer_replied?

* Add developer responsiveness badges

* Use rounded response rate for badges

* Move developer response rate into Developer model.

The developer response rate is stored in the `developers` table. This
value will be updated a configurable time after a new conversation is
intitiated by a Business.
This commit includes a migration to add the db column.
This commit includes a rake task (rake
backfills:developer_response_rate) to be run after the migration.

* rename 'time allowed' to 'grace period'

* Consolidate High Response Rate badges into a single badge.

 If a developer has a 90%+ response rate show a badge that says "High Response Rate".

* Move developer response rate updating to Messages::Notifications

* Remove migration check from backfills:developer_response_rate task

Let the job fail on its own when the migration hasn't been run.

* Improve developer response rate testing

Simplify testing now that we are storing the response rate as an actual percentage.
Additional tests to check that the grace period for replying to business contacts is being honored.

* Lint files

* Add a note to change persistance method later.

We are using update_column to avoid changing the updated_at
timestamp on the developer record which would trigger the rendering of a
'Recently Active' badge on the developer profile.
This should be changed to update! once we have implemented cacheing of 
the
badges in their own model.

* i18n-tasks normalize

* fix issue with seeds trying to schedule background jobs

We set the queue adapter to :async for UpdateDeveloperResponseRateJob to avoid a NotImplementedError when seeding Conversations and
set the config.developer_response_grace_period to zero so that the jobs are run immediately.

* Run UpdateDeveloperResponseRateJob after seeding Conversations

This ensures that the correct developer badges are seen in a freshly seeded development environment

* Fix developer response rate logic

This fix will ensure any conversations replied to within the grace period are also included
in the response rate calculation.

* Update CHANGELOG.md

Co-authored-by: Joe Masilotti <joe@masilotti.com>

* Remove unecessary calls to UpdateDeveloperResponseRateJob in seeds.

* Tidy up tests

* Add tests for Conversation#developer_replied?

* [skip ci] Update CHANGELOG

---------

Co-authored-by: Joe Masilotti <joe@masilotti.com>
Co-authored-by: Moses Gathuku <mosesgathuku95@gmail.com>